requirements.txt
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.12.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
Deprecated==1.2.18
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.104.1
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httpx==0.25.2
idna==3.10
limits==5.5.0
notion-client==2.0.0
openai==1.3.5
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGithub==2.1.1
PyJWT==2.10.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
python-telegram-bot==20.6
requests==2.31.0
six==1.17.0
slowapi==0.1.8
sniffio==1.3.1
soupsieve==2.8
starlette==0.27.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
watchdog==3.0.0
wrapt==1.17.3
